Complete reproduction: eval results + mechanistic figures + bug fixes
Results:
- Evasion: 100% TPR drop on deception and harmful probes
- Zero-shot transfer confirmed

Bug fixes:
- Fixed detach() in ActivationCache breaking grad flow
- Fixed bfloat16/float32 dtype mismatches
- Removed gradient_checkpointing incompatible with hooks
- Fixed double forward pass in compute_behavioral_loss

+++ evaluate.py
@@ -80,7 +80,7 @@ def evaluate_probe_evasion(
     neg_scores = []
     for act in neg_acts:
         with torch.no_grad():
-            score = probe(act.unsqueeze(0).to(device))
+            score = probe(act.unsqueeze(0).float().to(device))
             if score.dim() > 1:
                 score = score.mean(dim=-1)
             neg_scores.append(score.item())
@@ -97,7 +97,7 @@ def evaluate_probe_evasion(
     normal_scores = []
     for act in pos_acts_normal:
         with torch.no_grad():
-            score = probe(act.unsqueeze(0).to(device))
+            score = probe(act.unsqueeze(0).float().to(device))
             if score.dim() > 1:
                 score = score.mean(dim=-1)
             normal_scores.append(score.item())
@@ -117,7 +117,7 @@ def evaluate_probe_evasion(
     triggered_scores = []
     for act in pos_acts_triggered:
         with torch.no_grad():
-            score = probe(act.unsqueeze(0).to(device))
+            score = probe(act.unsqueeze(0).float().to(device))
             if score.dim() > 1:
                 score = score.mean(dim=-1)
             triggered_scores.append(score.item())
